Refactoring optimizer with new passes
