+++ dlk/python/dlk/core/optimizer.py
@@ -14,632 +14,401 @@
 # limitations under the License.
 # =============================================================================
 """Module of optimization passes."""
+import math
 import numpy as np
-from core.data_types import DataType, Float32, Float64, Int8, Int16, Int32, Int64, Uint8, Uint16, Uint32, Uint64, \
-    Bool, String, QUANTIZED_NOT_PACKED
-from core.graph import Graph, GraphRunner
-from core.operators import Add, AveragePool, BatchNormalization, Constant, Conv, Identity, Input, \
-    MaxPool, Operator, Output, Transpose, Quantizer, QTZ_binary_mean_scaling, QTZ_linear_mid_tread_half, \
-    Reshape, Softmax, Relu, Flatten, Dropout, Gemm, SpaceToDepth, QTZ_binary_channel_wise_mean_scaling, ConcatOnDepth,\
-    Maximum, DepthToSpace, Split, Variable
 
 from typing import Any, Dict, List, Optional, Set, cast
-from functools import reduce
-from enum import Enum
 
+from core.graph import Graph
+from core.graph_pattern_matching import GraphMatcher, Pattern, match_to_execution_list, NodeMatch
+from core.operators import Constant, Operator
+from core.data_types import Uint32, QUANTIZED_NOT_PACKED
+from typing import cast
+from collections import defaultdict
 from modules.packer import Packer
 
-NodeGroup = List[Operator]
 
+def pass_dot_graph(graph: Graph, filename) -> None:
 
-def node_is_add(node: Operator) -> bool:
-    return node.op_type == 'Add'
+    dot_script = 'digraph {'
 
+    code = {}
+    counter = 0
+    for node in graph.operators:
+        code[node.name] = counter
+        counter += 1
 
-def node_is_conv(node: Operator) -> bool:
-    return node.op_type == 'Conv'
+    for node in graph.operators:
 
+        shape = '-'
+        if node.shape:
+            shape = 'x'.join(str(x) for x in node.shape)
+        shape += '(' + node.dimension + ')'
 
-def node_is_concat(node: Operator) -> bool:
-    return node.op_type == 'ConcatV2'
+        dot_script += node.name + '[label="<f0> ' + format(code[node.name], '04X') + '| <f1> ' + \
+            node.op_type + '| <f2> ' + shape + '| <f3> ' + node.dtype.cpptype() + '" shape = "record"];'
+        for i in node.input_nodes:
+            dot_script += i.name + ' -> ' + node.name + ';'
 
+    dot_script += '}'
 
-def node_is_const(node: Operator) -> bool:
-    return node.op_type == 'Constant'
+    with open(filename, 'w') as f:
+        f.write(dot_script)
 
 
-def node_is_qconv(node: Operator) -> bool:
-    return node.op_type == 'Conv' and cast(Conv, node).is_quantized
+def pass_remove_identities(graph: Graph) -> None:
 
+    gm = GraphMatcher(graph)
 
-def node_is_input(node: Operator) -> bool:
-    return node.op_type == 'Input'
+    to_be_removed = list()
+    matches: List[NodeMatch] = list()
+    p = Pattern("Identity")
+    gm.get_op_type_matches(p, matches)
 
-
-def node_is_weight_quantizer(node: Operator) -> bool:
-    return (node.op_type == 'QTZ_binary_mean_scaling'
-            or node.op_type == 'QTZ_binary_channel_wise_mean_scaling')
-
-
-def node_is_activation_quantizer(node: Operator) -> bool:
-    return node.op_type == 'QTZ_linear_mid_tread_half'
-
-
-class NHWC_Transposer(GraphRunner):
-    """Transposer of all nodes to NHWC."""
-
-    def _get_permutation(self, dim: str) -> List[int]:
-        """Create a permutation from the source dimension."""
-        assert len(dim) == 4 and set(dim).issubset({'N', 'H', 'W', 'C', 'I', 'O'}), \
-            f'illegal dimension found: {dim}'
-
-        if set(dim) == set('HWIO'):
-            dim = dim.replace('I', 'C')
-            dim = dim.replace('O', 'N')
-
-        return list(map(lambda s: dim.index(s), 'NHWC'))
-
-    def _check_and_transpose(self, node: Operator) -> None:
-        perm = self._get_permutation(node.dimension)
-        node.transpose(perm)
-
-    def run_backward_input(self, node: Input, **kwargs: Any) -> None:
-        self._check_and_transpose(node)
-
-    def run_backward_constant(self, node: Constant, **kwargs: Any) -> None:
-        if node.ndims == 4 and set(node.dimension).issubset({'N', 'H', 'W', 'C', 'I', 'O'}):
-            self._check_and_transpose(node)
-
-    def run_backward_identity(self, node: Identity, **kwargs: Any) -> None:
-        if node.ndims == 4 and set(node.dimension).issubset({'N', 'H', 'W', 'C', 'I', 'O'}):
-            self._check_and_transpose(node)
-
-    def run_backward_QTZ_binary_mean_scaling(self, node: QTZ_binary_mean_scaling, **kwargs: Any) -> None:
-        self._check_and_transpose(node)
-
-    def run_backward_transpose(self, node: Transpose, **kwargs: Any) -> None:
-        raise NotImplementedError('Transposing Transpose operator is not supported yet.')
-
-    def run_backward_conv(self, node: Conv, **kwargs: Any) -> None:
-        self._check_and_transpose(node)
-
-    def run_backward_batch_normalization(self, node: BatchNormalization, **kwargs: Any) -> None:
-        self._check_and_transpose(node)
-
-    def run_backward_QTZ_linear_mid_tread_half(self, node: QTZ_linear_mid_tread_half, **kwargs: Any) -> None:
-        self._check_and_transpose(node)
-
-    def run_backward_max_pool(self, node: MaxPool, **kwargs: Any) -> None:
-        self._check_and_transpose(node)
-
-    def run_backward_average_pool(self, node: AveragePool, **kwargs: Any) -> None:
-        self._check_and_transpose(node)
-
-    def run_backward_SpaceToDepth(self, node: SpaceToDepth, **kwargs: Any) -> None:
-        self._check_and_transpose(node)
-
-    def run_backward_QTZ_binary_channel_wise_mean_scaling(
-            self,
-            node: QTZ_binary_channel_wise_mean_scaling,
-            **kwargs: Any) -> None:
-        self._check_and_transpose(node)
-
-    def run_backward_ConcatOnDepth(self, node: ConcatOnDepth, **kwargs: Any) -> None:
-        self._check_and_transpose(node)
-
-    def run_backward_Maximum(self, node: Maximum, **kwargs: Any) -> None:
-        self._check_and_transpose(node)
-
-    def run_backward_DepthToSpace(self, node: DepthToSpace, **kwargs: Any) -> None:
-        self._check_and_transpose(node)
-
-
-class PreComputeRunner(GraphRunner):
-    """Optimization class that does precomputation and pruning on the graph.
-
-    Fron a constant node, this object precomputes as far as possible, and
-    replaces all precomputed nodes with a newly defined constant node.
-
-    Additionally, in the hard-quantized mode, this object replaces a
-    weight-quantizer node and succesive Conv node with a QConv node, and
-    packs the weight.
-    """
-
-    _quantized_bitwidth = 1
-    _wordsize = 32
-
-    def __init__(self, graph: Graph, hard_quantized: bool = False) -> None:
-        """Set up internal varibles."""
-        self._precomp_dic: Dict[str, bool] = {}
-        self._nodes_removed: Set[Operator] = set()
-        self._hard_quantized = hard_quantized
-        self._quantizers: Dict[str, Quantizer] = {}  # the operator name and its quantizer
-        self._connected_convs: Dict[Operator, List[Conv]] = {}  # node name and its connected convolver
-
-        super().__init__(graph)
-
-    def initialize(self, **kwargs: Any) -> None:
-        qconvs: List[Conv] = kwargs['qconv']
-        self._connected_convs = {q: [q] for q in qconvs}
-
-    def finalize(self, **kwargs: Any) -> None:
-        """Remove all unused nodes from the graph."""
-        for n in self._nodes_removed:
-            self._graph.remove_op(n)
-
-    # 1st phase: check which conv the node connects
-
-    def run_backward_by_default(self, node: Operator, **kwargs: Any) -> None:
-        outputs = node.output_op_list
-
-        convs: List[Conv] = sum([self._connected_convs[out] for out in outputs if self._connected_convs.get(out)], [])
-        self._connected_convs[node] = convs
-
-    def run_backward_conv(self, node: Conv, **kwargs: Any) -> None:
-        pass  # do nothing, as all (quantized) conv node is already registered to self._connected_convs
-
-    # 2nd phase: precompute and prune
-
-    def _has_precompute_value(self, op: Operator) -> bool:
-        """Return True if the operator has precompute value."""
-        return self._precomp_dic[op.name]
-
-    def _is_prunable(self, op: Operator) -> bool:
-        """Return True if op can be prunable."""
-        return self._has_precompute_value(op) and op.op_type != 'Constant'
-
-    def _prune(self, node: Operator) -> None:
-        """Prune the node and its inputs."""
-        # prune inputs
-        for i in node.input_ops.values():
-            if i not in self._nodes_removed:
-                self._prune(i)
-
-        # prune itself
-        self._nodes_removed.add(node)
-
-    def _precompute_or_prune_inputs(self, node: Operator) -> None:
-        """Precompute itself or prune the input nodes.
-
-        If all input has precompute value, then make the node precompute.
-        Otherwise, all prunable input nodes are pruned and substituted with
-        a new constant node.
-        """
-        ops: List[Operator] = [node.input_ops[i] for i in node.input_names if node.input_ops.get(i)]
-        ops_have_precomp_values = list(map(lambda x: self._has_precompute_value(x), ops))
-        ops_are_prunable = list(map(lambda x: self._is_prunable(x), ops))
-        ops_are_in_quantized = list(map(lambda x: x.name in self._quantizers.keys(), ops))
-
-        # check which input node can be pruned
-        if reduce(lambda x, y: x and y, ops_have_precomp_values):  # all input has concrete values
-            node.run_forward()
-            self._precomp_dic[node.name] = True  # this node can be pruned
-            if reduce(lambda x, y: x or y, ops_are_in_quantized):  # some input operator to be quantized exists
-                quantizers = {op.name: self._quantizers[op.name] for op in ops if self._quantizers.get(op.name)}
-                if len(quantizers) > 1:
-                    ValueError(f'{node.name}: multiple quantized inputs with {node.op_type} are not supported.')
-                self._quantizers[node.name] = list(quantizers.values())[0]
-
-        else:
-            self._precomp_dic[node.name] = False
-
-            # prune input opetarots
-            for key, op in zip(node.input_names, ops):
-                if self._is_prunable(op):
-                    # get scaling factor if it is to be quantized but not in hard quantization mode
-                    scaling = 1 if self._quantizers.get(op.name) is None \
-                        else self._quantizers[op.name].scaling_factor
-
-                    extra_dims = tuple(np.ones((len(op.data.shape) - len(scaling.shape)), dtype=np.int32))
-                    scaling = scaling.reshape(scaling.shape + extra_dims)
-
-                    # creates new constant
-                    new_op = Constant(
-                        op.name + '_new',
-                        op.dtype,
-                        op.data * scaling,
-                        dimension_format=op.dimension
-                    )
-
-                    # replace and prune the old operators
-                    node.add_input(key, new_op)
-                    self._graph.add_op(new_op)
-                    self._prune(op)
-
-    def run_forward_by_default(self, node: Operator, **kwargs: Any) -> None:
-        self._precompute_or_prune_inputs(node)
-
-    def run_forward_input(self, node: Input, **kwargs: Any) -> None:
-        self._precomp_dic[node.name] = False
-
-    def run_forward_constant(self, node: Constant, **kwargs: Any) -> None:
-        self._precomp_dic[node.name] = True
-
-    def run_forward_identity(self, node: Identity, **kwargs: Any) -> None:
+    for m in matches:
         """skip all identity."""
-        in_op = node.input_ops['input']
-        out_ops = node.output_ops['output']
+        in_op = m.node.input_ops['input']
+        out_ops = m.node.output_ops['output']
         for out_op in out_ops:
             for k, v in out_op.input_ops.items():
-                if v == node:
+                if v == m.node:
                     # change the output's input to this identity's input
                     out_op.add_input(k, in_op)
                     # change the input's output to this identity's output
                     for k2, v2 in in_op.output_ops.items():
-                        if node in v2:
-                            v2.remove(node)
+                        if m.node in v2:
+                            v2.remove(m.node)
                             v2.append(out_op)
                             break
                     break
 
-    def run_forward_QTZ_binary_mean_scaling(self, node: QTZ_binary_mean_scaling, **kwargs: Any) -> None:
-        in_op = node.input_ops['input']
+        to_be_removed.append(m.node)
 
-        # if it can be precomputed
-        if self._has_precompute_value(in_op):
-            node.run_forward()
-            self._precomp_dic[node.name] = True  # this node can be pruned
-            self._quantizers[node.name] = node  # add itself as the quantizer
+    for op in to_be_removed:
+        graph.remove_op(op)
+
+
+def pass_transpose(graph: Graph) -> None:
+
+    gm = GraphMatcher(graph)
+
+    matches: List[NodeMatch] = list()
+    p = Pattern("*")
+    gm.get_op_type_matches(p, matches)
+
+    for m in matches:
+        dim = m.node.dimension
+        shape = m.node.shape
+        if len(shape) != 4 or len(dim) != 4 or not set(dim).issubset({'N', 'H', 'W', 'C', 'I', 'O'}):
+            continue
+
+        dim = dim.replace('I', 'C')
+        dim = dim.replace('O', 'N')
+
+        permutation = list(map(lambda s: dim.index(s), 'NHWC'))
+        m.node.transpose(permutation)
+
+
+def pass_precompute(graph: Graph, processed_nodes) -> bool:
+
+    gm = GraphMatcher(graph)
+
+    matches: List[NodeMatch] = list()
+    p = Pattern('*')
+    gm.get_op_type_matches(p, matches)
+
+    processed_before_precompute = len(processed_nodes)
+
+    for m in matches:
+        if m.node in processed_nodes:
+            continue
+
+        # We want operators with inputs
+        if not m.node.input_nodes:
+            continue
+
+        precomputable = True
+        for input_node in m.node.input_nodes:
+            if input_node.op_type != 'Constant':
+                precomputable = False
+
+        if not precomputable:
+            continue
+
+        processed_nodes += m.node.input_nodes
+        processed_nodes.append(m.node)
+
+        data = m.node.run_forward()
+
+        new_constant = Constant(
+            m.node.name + '_new',
+            m.node.dtype,
+            data,
+            dimension_format=m.node.dimension
+        )
+
+        graph.add_op(new_constant)
+
+        new_constant.add_outputs(m.node.output_ops)
+        for output_name, consumer_list in m.node.output_ops.items():
+            for consumer_node in consumer_list:
+                for input_name, input_node in consumer_node.input_ops.items():
+                    if input_node == m.node:
+                        consumer_node.add_input(input_name, new_constant)
+                        break
+    return len(processed_nodes) > processed_before_precompute
+
+
+def pass_propagate_quantization_details_into_conv(graph: Graph) -> None:
+
+    gm = GraphMatcher(graph)
+
+    matches: List[NodeMatch] = list()
+    p = Pattern('*')
+    gm.get_op_type_matches(p, matches)
+
+    qtypes = [
+        'QTZ_binary_mean_scaling',
+        'QTZ_linear_mid_tread_half',
+        'QTZ_binary_channel_wise_mean_scaling'
+    ]
+
+    quant_details = defaultdict(list)
+    for m in matches:
+        if not m.node.preserve_quantization:
+            quant_details[m.node.name] = []
+            continue
+
+        if m.node.op_type == 'Conv':
+            input_node = m.node.input_nodes[0]
+            weight_node = m.node.input_nodes[1]
+
+            m.node.a_quantizer = [input_node] if input_node.op_type in qtypes else quant_details[input_node.name]
+            m.node.quantizer = weight_node if weight_node.op_type in qtypes else quant_details[weight_node.name]
+
+            quant_details[m.node.name] = []
         else:
-            self._precomp_dic[node.name] = False
+            qtzs = []
+            for n in m.node.input_nodes:
+                if n.op_type in qtypes:
+                    qtzs.append(n)
+                else:
+                    for q in quant_details[n.name]:
+                        qtzs.append(q)
 
-    def run_forward_conv(self, node: Conv, **kwargs: Any) -> None:
-        ops: List[Operator] = [node.input_ops[i] for i in node.input_names if node.input_ops.get(i)]
+            quant_details[m.node.name] = qtzs if len(qtzs) == len(m.node.input_nodes) else []
+            # TODO: check if the quantizers use same n_bits
 
-        if self._hard_quantized and node in kwargs['qconv']:
-            # data is to be packed
-            ops_have_precomp_values = list(map(lambda x: self._has_precompute_value(x), ops))
-            ops_are_prunable = list(map(lambda x: self._is_prunable(x), ops))
 
-            # check which input node can be pruned
-            if reduce(lambda x, y: x and y, ops_have_precomp_values):  # all input has concrete values
-                node.run_forward()
-                self._precomp_dic[node.name] = True  # this node can be pruned
-                quantizers = {op.name: self._quantizers[op.name] for op in ops if self._quantizers.get(op.name)}
-                if len(quantizers) > 1:
-                    ValueError(f'{node.name}: multiple quantized inputs with {node.op_type} are not supported.')
-                self._quantizers[node.name] = list(quantizers.values())[0]
+def pass_compute_thresholds(graph: Graph) -> None:
 
-            else:   # an input (must be weight) is to be quantized and packed
-                self._precomp_dic[node.name] = False
-                node.is_quantized = True
-                packer = Packer(self._quantized_bitwidth, self._wordsize)
-                quantizers = {op.name: self._quantizers[op.name] for op in ops if self._quantizers.get(op.name)}
-                if len(quantizers) > 1:
-                    ValueError(f'{node.name}: multiple quantized inputs with {node.op_type} are not supported.')
-                node.quantizer = list(quantizers.values())[0]
+    gm = GraphMatcher(graph)
 
-                for key, op in zip(node.input_names, ops):
+    matches: List[NodeMatch] = list()
+    p = Pattern('QTZ_linear_mid_tread_half')
 
-                    if self._is_prunable(op):
-                        shape = op.shape
-                        op_data = node.quantizer.binarizer(op.data)
-                        data = packer.run(op_data.astype(np.float32), op.dimension)
-                        dtype = op.dtype
-                        new_op = Constant(
-                            op.name + '_new',
-                            dtype,
-                            data,
-                            packed=True,
-                            actual_shape=shape
-                        )
-                        node.add_input(key, new_op)
-                        self._graph.add_op(new_op)
-                        self._prune(op)
+    gm.get_op_type_matches(p, matches)
 
+    for m in matches:
+
+        p = [m.node]
+        while p[-1].op_type != 'Conv':
+            non_variable_input = [inode for inode in p[-1].input_nodes
+                                  if (not cast(Operator, inode).is_variable and inode.is_monotonic)
+                                  or inode.op_type == 'Conv']
+            if len(non_variable_input) != 1:
+                break
+            p.append(non_variable_input[-1])
+
+        if p[-1].op_type != 'Conv':
+            continue
+        quantizer_conv_output_node = p[0]
+        conv_node = p[-1]
+
+        # check if this is a quantized convolution
+        if not conv_node.quantizer or not conv_node.a_quantizer:
+            continue
+
+        quantizer_conv_weights = conv_node.quantizer
+        quantizer_conv_weights.run_forward_no_scaling_factor()
+        scaling_factor = quantizer_conv_weights.scaling_factor
+
+        # Getting the bit and max value
+        nbits = []
+        max_vs = []
+        for aqtz in conv_node.a_quantizer:
+            nbits.append(aqtz.nbit)
+            max_vs.append(aqtz.max_v)
+        if not (len(set(nbits)) == 1) and not (len(set(max_vs)) == 1):
+            raise ValueError(f'bits {nbits} or max values {max_vs} are not consistent')
         else:
-            self._precompute_or_prune_inputs(node)
+            nbit = nbits[0]
+            max_v = max_vs[0]
 
-    def run_forward_QTZ_binary_channel_wise_mean_scaling(
-            self,
-            node: QTZ_binary_channel_wise_mean_scaling,
-            **kwargs: Any) -> None:
-        in_op = node.input_ops['input']
+        n = 2 ** nbit - 1
+        ch = conv_node.channel
+        # assume that the threshold values will be a 13-bit signed integer
+        max_th_value = 2 ** 12 - 1
 
-        # if it can be precomputed
-        if self._has_precompute_value(in_op):
-            node.run_forward()
-            self._precomp_dic[node.name] = True  # this node can be pruned
-            self._quantizers[node.name] = node  # add itself as the quantizer
-        else:
-            self._precomp_dic[node.name] = False
+        # The threshold_table is numpy array that holds the threshold values for all channels
+        threshold_table = np.empty([ch, n + 1], dtype=np.int32)
 
+        # Compute threshold (t0, t1, t2)
+        th_val = [0.5 + i for i in range(n)]
+        for th_id, th_v in enumerate(th_val):
+            init_threshold = np.full(ch, th_v, dtype=np.float64)
 
-class DTypeChanger(GraphRunner):
-    """Optimization class that changes dypes.
+            # run calculation in reverse order: q -> bn -> scaling
+            trans_th = {'data': init_threshold}
+            for op in p[:-1]:
+                trans_th = op.de_run(**trans_th)
+            threshold = (trans_th['data'] * np.float64(n)) / (np.float64(max_v) * scaling_factor)
 
-    This runner must run before PrecomputeRunner.
-    """
+            for ch_id, th_per_ch in enumerate(threshold):
+                if quantizer_conv_weights.op_type == 'QTZ_binary_channel_wise_mean_scaling':
+                    threshold_table[ch_id, th_id] = int(math.floor(th_per_ch)) \
+                        if (scaling_factor[ch_id] < 0) ^ (ch_id in trans_th['nega_idx']) \
+                        else int(math.ceil(th_per_ch))
+                else:
+                    threshold_table[ch_id, th_id] = int(math.floor(th_per_ch)) \
+                        if (scaling_factor < 0) ^ (ch_id in trans_th['nega_idx']) \
+                        else int(math.ceil(th_per_ch))
 
-    class Path(Enum):
-        INPUT = 1,
-        WEIGHT = 2,
-        OTHER = 3
+        # take care of threshold values that are larger than 16-bit signed integer
+        threshold_table[abs(threshold_table) > max_th_value] = max_th_value
 
-    _packed_dtype = {Path.INPUT: QUANTIZED_NOT_PACKED(), Path.WEIGHT: Uint32(), Path.OTHER: Float32()}
-    _a_quantizers = {'QTZ_linear_mid_tread_half'}
-    _w_quantizers = {'QTZ_binary_mean_scaling', 'QTZ_binary_channel_wise_mean_scaling'}
-    _conv = {'Conv'}
+        for c in range(ch):
+            threshold_table[c, -1] = 1 \
+                if np.all(threshold_table[c, 1:-1] > threshold_table[c, :-2], axis=0) else -1
+            # Applying the magic number
+            if np.all(threshold_table[c, 1:-1] == threshold_table[c, :-2], axis=0):
+                threshold_table[c, -1] = 2
 
-    def __init__(self, graph: Graph) -> None:
-        """Set up internal varibles."""
-        self._output_convs: Dict[Operator, List[Conv]] = {}
-        self._packed_input_path: Dict[str, Any] = {}
+        # Put the thresholds into list
+        conv_node.thresholds = threshold_table.flatten().tolist()
 
-        super().__init__(graph, depth_first=False)
+        # Disconnect batchnorm and the quantizer
+        out_ops = quantizer_conv_output_node.output_ops['output']
+        for output_node in out_ops:
+            for input_name, input_node in output_node.input_ops.items():
+                if input_node == quantizer_conv_output_node:
+                    output_node.add_input(input_name, conv_node)
 
-    # 1st phase: check nodes which dtype must be changed
-
-    def _check_dtype_state(self, node: Operator) -> None:
-        """checks the state of each node regarding dtype.
-
-        - whether the node is after conv and before activation quantizer
-        - whether the node is after activation and before conv
-        """
-        outputs = node.output_op_list
-        convs: List[Conv] = sum([self._output_convs[out] for out in outputs if self._output_convs.get(out) is not None],
-                                [])
-
-        # determine the path of node is input or weight or others
-        path = self.Path.WEIGHT
-        for out in outputs:
-            p = self._packed_input_path[out.name] if out.op_type not in self._conv \
-                else self.Path.INPUT if node == out.input_ops['X'] \
-                else self.Path.WEIGHT
-            if path == self.Path.WEIGHT:
-                path = p
-            elif path == p:
-                pass
-            else:  # output have different paths
-                ValueError('multiple outputs must have the same kind of paths.')
-
-        is_not_before_a_quantizer = reduce(lambda x, y: x and y,
-                                           [out.op_type not in self._a_quantizers for out in outputs])
-        if convs and is_not_before_a_quantizer:
-            self._output_convs[node] = convs
-
-        self._packed_input_path[node.name] = path
-
-    def run_backward_by_default(self, node: Operator, **kwargs: Any) -> None:
-        self._check_dtype_state(node)
-
-    def run_backward_output(self, node: Output, **kwargs: Any) -> None:
-        self._packed_input_path[node.name] = self.Path.OTHER
-
-    def run_backward_conv(self, node: Conv, **kwargs: Any) -> None:
-        self._output_convs[node] = [node]
-
-    # 2nd phase: change data type
-
-    def turn(self, **kwargs: Any) -> None:
-        """Set up qconv list"""
-        output_convs: List[Conv] = sum(list(self._output_convs.values()), [])
-        for conv in output_convs:
-            # get all ascendants of conv
-            ascendants = [k for k in self._output_convs.keys() if conv in self._output_convs[k]]
-
-            # whether some weight quantizer is in ascendants
-            wqtz_in_asc = reduce(lambda x, y: x or y,
-                                 list(map(lambda n: n.op_type in self._w_quantizers, ascendants)))
-            # whether some activation quantizer is in ascendants
-            aqtz_in_asc = reduce(lambda x, y: x or y,
-                                 list(map(lambda n: n.op_type in self._a_quantizers, ascendants)))
-            # if both, add conv to the list
-            if wqtz_in_asc and aqtz_in_asc:
-                kwargs['qconv'].add(conv)
-
-    def _set_dtype(self, node: Operator, qconv: List[Conv]) -> None:
-        def before_qconv() -> bool:
-            """Return if the node is before a quantized convolver"""
-            convs: List[Conv] = self._output_convs[node] if self._output_convs.get(node) else []
-            # consistency check
-            is_qconv: List[bool] = list(map(lambda x: x in qconv, convs))
-            all_is_qconv = reduce(lambda x, y: x and y, is_qconv, True)
-            some_is_qconv = reduce(lambda x, y: x or y, is_qconv, False)
-            assert convs == [] or (all_is_qconv == some_is_qconv), \
-                f'{node.name} connects to both of a quantized convolver and non-quantized one.'
-
-            return convs != [] and all_is_qconv
-
-        def get_dtype() -> Optional[DataType]:
-            """Return dtype along with which path the node is on: 'input' or 'weight' of a conv"""
-            path = self._packed_input_path.get(node.name)
-            return self._packed_dtype[path] if path is not None else None
-
-        dtype = get_dtype()
-        conv = self._output_convs.get(node)
-        if dtype is not None and before_qconv():
-            node.dtype = dtype
-
-    def run_forward_by_default(self, node: Operator, **kwargs: Any) -> None:
-        self._set_dtype(node, kwargs['qconv'])
+        conv_node.remove_output('Y')
+        conv_node.add_outputs({'Y': out_ops})
 
 
-class ApplyThresholdSkipping(GraphRunner):
-    """Optimization class that perform threshold skipping.
+def pass_pack_weights(graph: Graph) -> None:
 
-    This runner perform threshold skipping with BFS for DLK graph.
-    Run graphrunner backward to acquire graph info, and run forward
-    to compute the thresholds skip batchnorm and activation quantizer
-    with thresholding function.
-    """
+    gm = GraphMatcher(graph)
 
-    def __init__(self, graph: Graph) -> None:
-        self._aqtz_aqtz: Dict[Operator, List[Operator]] = {}
-        self._qconv_qconv: Dict[Conv, List] = {}
-        super().__init__(graph, depth_first=False)
+    quantization_types = [
+        'QTZ_binary_mean_scaling',
+        'QTZ_linear_mid_tread_half',
+        'QTZ_binary_channel_wise_mean_scaling'
+    ]
 
-    def _apply_threshold_skipping(self, op_lst: List[Operator]) -> None:
-        """Performs Conv thresholds computation and skipping."""
+    matches: List[NodeMatch] = list()
+    p = Pattern('Conv')
 
-        transitions: Dict[int, Operator] = {}
-        start, finish = [None, None]
-        for idx, op in enumerate(op_lst):
-            if node_is_qconv(op):
-                start = cast(Conv, op)
-            elif node_is_activation_quantizer(op):
-                finish = op
-                transitions[idx] = op
-            else:
-                transitions[idx] = op
+    gm.get_op_type_matches(p, matches)
 
-        if start is not None and finish is not None:
+    # TODO: pass proper parameters
+    packer = Packer(1, 32)
 
-            # def linear_qtz2float(x: np.ndarray, n_value: int, max_value: float) -> np.ndarray:
-            #     real_x = x / np.float64(n_value) * np.float64(max_value)
-            #     return real_x.astype(np.float64)
+    for m in matches:
+        conv_node = m.node
 
-            # Step 1: Compute thresholds for Convolution operators
-            # aqtzer = cast(Quantizer, start.a_quantizer[0])  # Activation Quantizers should all have the same bits
-            # bit = aqtzer.nbit
-            # max_v = aqtzer.max_v
-            # if bit is None or max_v is None:
-            #     ValueError(f'activation quantizer of node {start.name} has bit or max value of None')
+        # check if this is a quantized convolution
+        if not conv_node.quantizer or not conv_node.a_quantizer:
+            continue
 
-            # n = 2 ** bit - 1
-            # ch = start.channel
-            # lch = start.input_ops['X'].channel
-            # k = start.kernel_height * start.kernel_width * lch * n
-            # qtzer = cast(Quantizer, start.quantizer)
-            # conv_results = [x for x in range(-k, k + 1, 1)]
-            # th_tmp = np.empty([ch, n + 1], dtype=np.int32)
-            # v_now = dict.fromkeys([x for x in range(ch)], 0)
-            # th_now = 0
-            # val_neg_flag = -1
-            # val_pos_flag = 1
-            # all_transdata: Dict[int, Dict[str, Any]] = {}
+        weight_quantizer = conv_node.quantizer
+        if weight_quantizer.op_type not in quantization_types:
+            continue
 
-            # Step 1-1: initalize thresholds
-            # for conv_res in conv_results:
-            #     conv_out = np.full(ch, conv_res, dtype=np.float64)
-            #     conv_out *= qtzer.scaling_factor if qtzer.scaling_factor is not None \
-            #         else ValueError(f'oops Quantizer of node {start.name} has scaling factor of None')
-            #
-            #     conv_data = linear_qtz2float(conv_out, n, max_v)
-            #
-            #     trans_data: Dict[str, Any] = {'data': conv_data}
-            #     for idx, op in sorted(transitions.items(), reverse=True):
-            #         trans_data = op.run(**trans_data)
-            #
-            #     for depth in range(ch):
-            #         init = -k if depth in trans_data['nega_idx'] else k
-            #         th_tmp[depth, :] = init
-            #
-            #     all_transdata[conv_res] = trans_data
+        # Quantize the weights
+        weight_quantizer.run_forward()
+        op_data = weight_quantizer.binarizer(weight_quantizer.data)
+        data = packer.run(op_data.astype(np.float32), weight_quantizer.dimension)
 
-            # Step 1-2: update thresholds
-            # for conv_res in conv_results:
-            #     trans_data = all_transdata[conv_res]
-            #     qtz_out = trans_data['data']
-            #     qtz_mu = np.mean(qtz_out)
-            #     if qtz_mu != th_now:
-            #         for depth in range(ch):
-            #             is_negative = depth in trans_data['nega_idx']
-            #             if v_now.get(depth) != qtz_out[depth]:
-            #                 if is_negative:
-            #                     th_tmp[depth, abs(n - qtz_out[depth] - 1)] = conv_res
-            #                 else:
-            #                     th_tmp[depth, qtz_out[depth] - 1] = conv_res
-            #                 v_now[depth] = qtz_out[depth]
-            #             th_tmp[depth, n] = -1 if is_negative else 1
-            #     for depth in range(ch):
-            #         constant = reduce(lambda x, y: x and y,
-            #                           [th_tmp[depth, i] == th_tmp[depth, i + 1] for i in range(n - 1)])
-            #         th_tmp[depth, n] = qtz_out[depth] + 2 if constant else th_tmp[depth, n]
-            #         # note: 2 above is a magic number. the result value must not be 1 nor -1.
-            #     th_now = qtz_mu
+        quantized_constant = Constant(
+            weight_quantizer.name + '_new',
+            Uint32(),
+            data,
+            packed=True,
+            actual_shape=weight_quantizer.shape
+        )
 
-            # start.thresholds = th_tmp.flatten().tolist()
+        graph.add_op(quantized_constant)
 
-            # Step 2: Skipping unused operators, e.g. batch normalization, linear activation quantizer
-            pass
-            # if start.has_thresholds:
-            #     if start.dtype is not finish.dtype:
-            #         start.dtype = finish.dtype
-            #     for consumers in finish.output_ops.values():
-            #         for consumer in consumers:
-            #             for idex, y in start.output_ops.items():
-            #                 if not bool(set(consumers) & set(y)):
-            #                     start.remove_output(idex)
-            #                 start.add_output(idex, consumer)
-            #
-            #             for indent, v in consumer.input_ops.items():
-            #                 if v == finish:
-            #                     consumer.add_input(indent, start)
-            #                     break
-        else:
-            pass
-
-    def _makeup_skippable(self, node: Operator) -> None:
-        outputs = node.output_op_list
-        for out_op in outputs:
-            for start, lst in self._aqtz_aqtz.items():
-                if out_op in lst:
-                    self._aqtz_aqtz[start].append(node)
-
-    def _makeup_aqtz(self, node: Operator) -> None:
-        outputs = node.output_op_list
-        for out_op in outputs:
-            for start, lst in self._qconv_qconv.items():
-                if out_op in lst:
-                    self._qconv_qconv[start].append(node)
-
-    def run_backward_QTZ_linear_mid_tread_half(self, node: QTZ_linear_mid_tread_half, **kwargs: Any) -> None:
-        self._aqtz_aqtz[node] = [node]
-        self._makeup_aqtz(node)
-
-    def run_backward_by_default(self, node: Operator, **kwargs: Any) -> None:
-        if node.is_monotonic and not node_is_conv(node):
-            self._makeup_skippable(node)
-        self._makeup_aqtz(node)
-
-    def run_backward_conv(self, node: Conv, **kwargs: Any) -> None:
-        self._makeup_skippable(node)
-        if node_is_qconv(node):
-            self._qconv_qconv[node] = [node]
-
-    def run_forward_conv(self, node: Conv, **kwargs: Any) -> None:
-        bits: List[int] = []
-        aqtzers: List[Quantizer] = []
-        if node_is_qconv(node):
-            for x in self._qconv_qconv[node]:
-                if node_is_activation_quantizer(x):
-                    bits.append(x.nbit)
-                    aqtzers.append(x)
-
-        if not (len(set(bits)) == 1):
-            ValueError('Values are not consistent')
-        else:
-            node.a_quantizer = aqtzers
-
-    def run_forward_QTZ_linear_mid_tread_half(self, node: QTZ_linear_mid_tread_half, **kwargs: Any) -> None:
-        self._apply_threshold_skipping(self._aqtz_aqtz[node])
+        quantized_constant.add_outputs(weight_quantizer.output_ops)
+        for output_name, consumer_list in weight_quantizer.output_ops.items():
+            for consumer_node in consumer_list:
+                for input_name, input_node in consumer_node.input_ops.items():
+                    if input_node == weight_quantizer:
+                        consumer_node.add_input(input_name, quantized_constant)
+                        break
 
 
-class Optimizer(object):
-    """Class of optimization classes."""
+def pass_quantize_convolutions(graph: Graph) -> None:
 
-    def transpose_NHWC(self, graph: Graph) -> Graph:
-        runner = NHWC_Transposer(graph)
-        kwargs: Dict[str, Any] = {}
-        runner.run(**kwargs)
-        return graph
+    gm = GraphMatcher(graph)
 
-    def precompute(self, graph: Graph, hard_quantized: bool = False) -> Graph:
-        runner1 = DTypeChanger(graph)
-        runner2 = PreComputeRunner(graph, hard_quantized=hard_quantized)
+    matches: List[NodeMatch] = list()
+    p = Pattern('Conv')
+    gm.get_op_type_matches(p, matches)
 
-        kwargs: Dict[str, Set[Conv]] = {'qconv': set()}
+    for m in matches:
+        conv_node = m.node
 
-        # run
-        if hard_quantized:
-            runner1.run(**kwargs)
-        runner2.run(**kwargs)
+        # check if this is a quantized convolution
+        if not conv_node.quantizer or not conv_node.a_quantizer:
+            continue
 
-        return graph
+        # Mark as quantized convolution
+        conv_node.is_quantized = True
 
-    def threshold_skipping(self, graph: Graph) -> Graph:
-        runner1 = ApplyThresholdSkipping(graph)
-        kwargs: Dict[str, Any] = {}
-        runner1.run(**kwargs)
-        return graph
+        # change the output data type of the convolution if thresholds are available
+        if conv_node.has_thresholds:
+            conv_node.dtype = QUANTIZED_NOT_PACKED
+
+        # change the output data type of the quantizers
+        conv_node.quantizer.dtype = Uint32
+        for qtz in conv_node.a_quantizer:
+            qtz.dtype = QUANTIZED_NOT_PACKED
+
+
+def pass_propagate_datatypes(graph) -> None:
+
+    gm = GraphMatcher(graph)
+
+    matches: List[NodeMatch] = list()
+    p = Pattern('*')
+    gm.get_op_type_matches(p, matches)
+
+    for m in matches:
+        if m.node.op_type != 'Conv' and m.node.preserve_quantization:
+            m.node.dtype = m.node.input_nodes[0].dtype
+
+
+def pass_propagate_output_type_backward(graph: Graph) -> None:
+
+    gm = GraphMatcher(graph)
+
+    matches: List[NodeMatch] = list()
+    p = Pattern('*')
+
+    gm.get_op_type_matches(p, matches)
+
+    def find_input(node, otype):
+        for n in node.input_nodes:
+            if n.op_type == 'Conv' and n.is_quantized:
+                n.dtype = otype
+                return
+            find_input(n, otype)
+
+    output_node = matches[-1].node
+
+    output_type = output_node.dtype
+    find_input(output_node, output_type)
